New attribute for misc. annotations
Added a new attribute for miscellaneous annotations, which are echoed into the XML for the sake of external tools that process the XML.

Fixed an issue where integral types used exclusively within highly specific contexts (e.g. union members-of-members) could be missed in the XML output. It's an ugly, brute-force fix, but this plug-in is a means to an end, and right now I'm working more on the end than on the means.