XML generation refactored
We now permit stat-categories and misc-annotations on types, and output them for struct and union types in our XML. If they are applied to a type, then they are not output for the type's instances.

The logic for outputting an instruction tree as XML has been split away from `report_generator`, to keep that class more manageable.

In general, the process for outputting XML should produce more consistent results and should be more maintainable.